Separate Tile and GameBoard into different files
