+++ src/com/bulenkov/game2048/GameBoard.java
@@ -1,0 +1,234 @@
+/*
+ * Copyright 1998-2014 Konstantin Bulenkov http://bulenkov.com/about
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.bulenkov.game2048;
+
+import java.util.ArrayList;
+import java.util.LinkedList;
+import java.util.List;
+
+/**
+ * @author Konstantin Bulenkov
+ */
+public class GameBoard {
+  Tile[] tiles;
+  boolean win;
+  int score;
+
+  public GameBoard() {
+    resetGame();
+  }
+
+  public void resetGame() {
+    score = 0;
+    win = false;
+    tiles = new Tile[4 * 4];
+    for (int i = 0; i < tiles.length; i++) {
+      tiles[i] = new Tile();
+    }
+    addTile();
+    addTile();
+  }
+
+  public boolean getWin() {
+    return win;
+  }
+  
+  public boolean getLose() {
+    return !win && !canMove();
+  }
+  
+  public int getScore() {
+    return score;
+  }
+  
+  public Tile getTile(int x, int y) {
+    return tiles[x + y * 4];
+  }
+  
+  public void left() {
+    boolean needAddTile = false;
+    for (int i = 0; i < 4; i++) {
+      Tile[] line = getLine(i);
+      Tile[] merged = mergeLine(moveLine(line));
+      setLine(i, merged);
+      if (!needAddTile && !compare(line, merged)) {
+        needAddTile = true;
+      }
+    }
+
+    if (needAddTile) {
+      addTile();
+    }
+  }
+
+  public void right() {
+    tiles = rotate(180);
+    left();
+    tiles = rotate(180);
+  }
+
+  public void up() {
+    tiles = rotate(270);
+    left();
+    tiles = rotate(90);
+  }
+
+  public void down() {
+    tiles = rotate(90);
+    left();
+    tiles = rotate(270);
+  }
+
+  public void addTile() {
+    List<Tile> list = availableSpace();
+    if (!availableSpace().isEmpty()) {
+      int index = (int) (Math.random() * list.size()) % list.size();
+      Tile emptyTime = list.get(index);
+      emptyTime.value = Math.random() < 0.9 ? 2 : 4;
+    }
+  }
+
+  public boolean canMove() {
+    if (!isFull()) {
+      return true;
+    }
+    for (int x = 0; x < 4; x++) {
+      for (int y = 0; y < 4; y++) {
+        Tile t = tileAt(x, y);
+        if ((x < 3 && t.value == tileAt(x + 1, y).value)
+          || ((y < 3) && t.value == tileAt(x, y + 1).value)) {
+          return true;
+        }
+      }
+    }
+    return false;
+  }
+
+  private Tile[] getLine(int index) {
+    Tile[] result = new Tile[4];
+    for (int i = 0; i < 4; i++) {
+      result[i] = tileAt(i, index);
+    }
+    return result;
+  }
+
+  private void setLine(int index, Tile[] re) {
+    System.arraycopy(re, 0, tiles, index * 4, 4);
+  }
+
+  private Tile[] moveLine(Tile[] oldLine) {
+    LinkedList<Tile> l = new LinkedList<Tile>();
+    for (int i = 0; i < 4; i++) {
+      if (!oldLine[i].isEmpty())
+        l.addLast(oldLine[i]);
+    }
+    if (l.size() == 0) {
+      return oldLine;
+    } else {
+      Tile[] newLine = new Tile[4];
+      ensureSize(l, 4);
+      for (int i = 0; i < 4; i++) {
+        newLine[i] = l.removeFirst();
+      }
+      return newLine;
+    }
+  }
+
+  private Tile[] mergeLine(Tile[] oldLine) {
+    LinkedList<Tile> list = new LinkedList<Tile>();
+    for (int i = 0; i < 4 && !oldLine[i].isEmpty(); i++) {
+      int num = oldLine[i].value;
+      if (i < 3 && oldLine[i].value == oldLine[i + 1].value) {
+        num *= 2;
+        score += num;
+        int ourTarget = 2048;
+        if (num == ourTarget) {
+          win = true;
+        }
+        i++;
+      }
+      list.add(new Tile(num));
+    }
+    if (list.size() == 0) {
+      return oldLine;
+    } else {
+      ensureSize(list, 4);
+      return list.toArray(new Tile[4]);
+    }
+  }
+
+  private static void ensureSize(java.util.List<Tile> l, int s) {
+    while (l.size() != s) {
+      l.add(new Tile());
+    }
+  }
+
+  private List<Tile> availableSpace() {
+    final List<Tile> list = new ArrayList<Tile>(16);
+    for (Tile t : tiles) {
+      if (t.isEmpty()) {
+        list.add(t);
+      }
+    }
+    return list;
+  }
+
+  private boolean isFull() {
+    return availableSpace().size() == 0;
+  }
+
+  private boolean compare(Tile[] line1, Tile[] line2) {
+    if (line1 == line2) {
+      return true;
+    } else if (line1.length != line2.length) {
+      return false;
+    }
+
+    for (int i = 0; i < line1.length; i++) {
+      if (line1[i].value != line2[i].value) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private Tile[] rotate(int angle) {
+    Tile[] newTiles = new Tile[4 * 4];
+    int offsetX = 3, offsetY = 3;
+    if (angle == 90) {
+      offsetY = 0;
+    } else if (angle == 270) {
+      offsetX = 0;
+    }
+
+    double rad = Math.toRadians(angle);
+    int cos = (int) Math.cos(rad);
+    int sin = (int) Math.sin(rad);
+    for (int x = 0; x < 4; x++) {
+      for (int y = 0; y < 4; y++) {
+        int newX = (x * cos) - (y * sin) + offsetX;
+        int newY = (x * sin) + (y * cos) + offsetY;
+        newTiles[(newX) + (newY) * 4] = tileAt(x, y);
+      }
+    }
+    return newTiles;
+  }
+
+  private Tile tileAt(int x, int y) {
+    return tiles[x + y * 4];
+  }
+}
